test p

+++ DayOne/ConsoleApplication1/ConsoleApplication1/MonHeader.hpp
@@ -2,9 +2,9 @@
 
 
 struct Vec3 {
-	float x;
-	float y;
-	float z;
+	int x;
+	int y;
+	int	z;
 };
 
 /*typedef struct Vec3{
